Revise unit tests around IPv6
