+++ dhcp-clients-webapp-backend/pkg/uibackend/types.go
@@ -1,10 +1,8 @@
 package uibackend
 
 import (
-	"dhcp-clients-webapp-backend/pkg/ippool"
 	"dhcp-clients-webapp-backend/pkg/trackerdb"
 	"encoding/json"
-	"fmt"
 	htmltemplate "html/template"
 	"net"
 	"net/netip"
@@ -26,237 +24,6 @@ type IpAddressReservation struct {
 	Mac  net.HardwareAddr
 	IP   netip.Addr
 	Link *texttemplate.Template // maybe nil
-}
-
-// IpNetworkInfo contains all details about a network attached to the addon, as specified in the config file
-type IpNetworkInfo struct {
-	Interface string
-	Start     net.IP
-	End       net.IP
-	Gateway   net.IP
-	Netmask   net.IPMask
-}
-
-func (nw IpNetworkInfo) HasValidIPs() bool {
-	if nw.Start.IsPrivate() &&
-		nw.End.IsPrivate() &&
-		nw.Gateway.IsPrivate() {
-
-		// net.IPMask.Size() returns 0,0 in case of invalid netmasks
-		if ones, zeros := nw.Netmask.Size(); ones != 0 && zeros != 0 {
-			// all private IPs and mask is valid: ok
-
-			// check start/end IPs are in the same network
-
-			theNetwork := net.IPNet{IP: nw.Start, Mask: nw.Netmask}
-			return theNetwork.Contains(nw.End)
-		}
-	}
-	return false
-}
-
-func (nw IpNetworkInfo) HasValidGateway() bool {
-	// Ensure IPs are of the same family (IPv4 or IPv6)
-	theNetwork := net.IPNet{IP: nw.Start, Mask: nw.Netmask}
-	return theNetwork.Contains(nw.Gateway)
-}
-
-func (nw IpNetworkInfo) String() string {
-	return fmt.Sprintf("Interface: %s, Start: %s, End: %s, Gateway: %s, Netmask: %s",
-		nw.Interface, nw.Start, nw.End, nw.Gateway, nw.Netmask)
-}
-
-// AddonConfig contains the configuration provided by the user to the Home Assistant addon
-// in the HomeAssistant YAML editor
-type AddonConfig struct {
-	// Static IP addresses, as read from the configuration
-	ipAddressReservationsByIP  map[netip.Addr]IpAddressReservation
-	ipAddressReservationsByMAC map[string]IpAddressReservation
-
-	// DHCP client friendly names, as read from the configuration
-	// The key of this map is the MAC address formatted as string (since net.HardwareAddr is not a valid map key type)
-	friendlyNames map[string]DhcpClientFriendlyName
-
-	// Multiple IP ranges all together form the DHCP pool
-	dhcpPool   ippool.Pool     // this type provide the Size() and Contains() methods
-	dhcpRanges []IpNetworkInfo // this type stores additional metadata for each network
-
-	// Log this backend activities?
-	logDHCP  bool
-	logWebUI bool
-
-	webUIPort int
-
-	// Lease times
-	defaultLease            string
-	addressReservationLease string
-
-	// DNS
-	dnsEnable bool
-	dnsDomain string
-	dnsPort   int
-}
-
-// UnmarshalJSON reads the configuration of this Home Assistant addon and converts it
-// into maps and slices that get stored into the UIBackend instance
-func (b *AddonConfig) UnmarshalJSON(data []byte) error {
-
-	// JSON structure.
-	// This must be updated every time the config.yaml of the addon is changed;
-	// however this structure contains only fields that are relevant to the
-	// UI backend behavior. In other words the addon config.yaml might contain
-	// more settings than those listed here.
-	var cfg struct {
-		DhcpIpAddressReservations []struct {
-			Name string `json:"name"`
-			Mac  string `json:"mac"`
-			IP   string `json:"ip"`
-			Link string `json:"link"`
-		} `json:"dhcp_ip_address_reservations"`
-
-		DhcpClientsFriendlyNames []struct {
-			Name string `json:"name"`
-			Mac  string `json:"mac"`
-			Link string `json:"link"`
-		} `json:"dhcp_clients_friendly_names"`
-
-		DhcpServer struct {
-			LogDHCP                 bool   `json:"log_requests"`
-			DefaultLease            string `json:"default_lease"`
-			AddressReservationLease string `json:"address_reservation_lease"`
-		} `json:"dhcp_server"`
-
-		DhcpPool []struct {
-			Interface string `json:"interface"`
-			Start     string `json:"start"`
-			End       string `json:"end"`
-			Gateway   string `json:"gateway"`
-			Netmask   string `json:"netmask"`
-		} `json:"dhcp_pools"`
-
-		DnsServer struct {
-			Enable    bool   `json:"enable"`
-			DnsDomain string `json:"dns_domain"`
-			Port      int    `json:"port"`
-		} `json:"dns_server"`
-
-		WebUI struct {
-			Log  bool `json:"log_activity"`
-			Port int  `json:"port"`
-		} `json:"web_ui"`
-	}
-	err := json.Unmarshal(data, &cfg)
-	if err != nil {
-		return err
-	}
-
-	// convert DHCP IP addresses (strings) to iprange.Pool == []iprange.Range
-	for _, r := range cfg.DhcpPool {
-		dhcpR := ippool.NewRangeFromString(r.Start, r.End)
-		if !dhcpR.IsValid() {
-			return fmt.Errorf("invalid DHCP range %s-%s found in addon config file", r.Start, r.End)
-		}
-
-		// create also the IpNetworkInfo obj associated:
-		ipNetInfo := IpNetworkInfo{
-			Interface: r.Interface,
-			Start:     dhcpR.Start,
-			End:       dhcpR.End,
-			Gateway:   net.ParseIP(r.Gateway),
-		}
-
-		m := net.ParseIP(r.Netmask)
-		if m.To4() != nil {
-			ipNetInfo.Netmask = net.IPMask(m.To4())
-		} else {
-			ipNetInfo.Netmask = net.IPMask(m.To16())
-		}
-
-		// check network definition
-		if !ipNetInfo.HasValidIPs() {
-			return fmt.Errorf("invalid DHCP network/range [%s] found in addon config file: the IP addresses should define a coherent network (private IPs only, start and end IPs within the same network)", ipNetInfo.String())
-		}
-		if !ipNetInfo.HasValidGateway() {
-			return fmt.Errorf("invalid DHCP network/range [%s] found in addon config file: the gateway must be an IP address within the network", ipNetInfo.String())
-		}
-
-		// all good: store the info
-		b.dhcpPool.Ranges = append(b.dhcpPool.Ranges, dhcpR)
-		b.dhcpRanges = append(b.dhcpRanges, ipNetInfo)
-	}
-
-	// ensure we have a valid port for web UI
-	if cfg.WebUI.Port <= 0 || cfg.WebUI.Port > 32768 {
-		return fmt.Errorf("invalid web UI port number: %d", cfg.WebUI.Port)
-	}
-
-	// convert IP address reservations to a map indexed by IP
-	for _, r := range cfg.DhcpIpAddressReservations {
-		ipAddr, err := netip.ParseAddr(r.IP)
-		if err != nil {
-			return fmt.Errorf("invalid IP address found inside 'ip_address_reservations': %s", r.IP)
-		}
-		macAddr, err := net.ParseMAC(r.Mac)
-		if err != nil {
-			return fmt.Errorf("invalid MAC address found inside 'ip_address_reservations': %s", r.Mac)
-		}
-
-		var linkTemplate *texttemplate.Template
-		if r.Link != "" {
-			linkTemplate, err = texttemplate.New("linkTemplate").Parse(r.Link)
-			if err != nil {
-				return fmt.Errorf("invalid golang template found inside 'link': %s", r.Link)
-			}
-		}
-
-		// normalize the IP and MAC address format (e.g. to lowercase)
-		r.IP = ipAddr.String()
-		r.Mac = macAddr.String()
-
-		ipReservation := IpAddressReservation{
-			Name: r.Name,
-			Mac:  macAddr,
-			IP:   ipAddr,
-			Link: linkTemplate,
-		}
-
-		b.ipAddressReservationsByIP[ipAddr] = ipReservation
-		b.ipAddressReservationsByMAC[macAddr.String()] = ipReservation
-	}
-
-	// convert friendly names to a map of DhcpClientFriendlyName instances indexed by MAC address
-	for _, client := range cfg.DhcpClientsFriendlyNames {
-		macAddr, err := net.ParseMAC(client.Mac)
-		if err != nil {
-			return fmt.Errorf("invalid MAC address found inside 'dhcp_clients_friendly_names': %s", client.Mac)
-		}
-
-		var linkTemplate *texttemplate.Template
-		if client.Link != "" {
-			linkTemplate, err = texttemplate.New("linkTemplate").Parse(client.Link)
-			if err != nil {
-				return fmt.Errorf("invalid golang template found inside 'link': %s", client.Link)
-			}
-		}
-
-		b.friendlyNames[macAddr.String()] = DhcpClientFriendlyName{
-			MacAddress:   macAddr,
-			FriendlyName: client.Name,
-			Link:         linkTemplate,
-		}
-	}
-
-	// copy basic settings
-	b.logDHCP = cfg.DhcpServer.LogDHCP
-	b.logWebUI = cfg.WebUI.Log
-	b.webUIPort = cfg.WebUI.Port
-	b.defaultLease = cfg.DhcpServer.DefaultLease
-	b.addressReservationLease = cfg.DhcpServer.AddressReservationLease
-	b.dnsEnable = cfg.DnsServer.Enable
-	b.dnsDomain = cfg.DnsServer.DnsDomain
-	b.dnsPort = cfg.DnsServer.Port
-
-	return nil
 }
 
 // DhcpClientData holds all the information the backend has about a particular DHCP client,
